splitted class TestAppendCommon to file test_append_common.py
